adiciona, permissoes de usuario com policy

+++ app/Http/Controllers/ProjectsController.php
@@ -41,6 +41,7 @@ class ProjectsController extends Controller
 
     public function show(Project $project)
     {
+        $this->authorize('update', $project);
         return view('projects.show', compact('project'));
     }
 
@@ -54,6 +55,7 @@ class ProjectsController extends Controller
     public function update(Request $request, Project $project)
     {
         //$project->update(request(['title', 'description']));
+        $this->authorize('update', $project);
         $attributes = request()->validate([
             'title' => 'required|min:3|max:255',
             'description' => 'required|min:3'
@@ -64,6 +66,7 @@ class ProjectsController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('update', $project);
         $project->delete();
         return redirect('/projects');
     }
